requirements.txt
Flask
Flask-Cors
mysql-connector-python
python-dotenv
werkzeug
